GS-2288 Update redirect module to provide canonical url redirect
This commit sets the version of the redirect module to a known
commit and applies a patch that includes functionality that was
in the (now deprecated) globalredirect module.

It ensures that a page is always served from the URL that is in
its rel='canonical' tag (usually its alias). This prevents content
being duplicated under node/<nid> and /<alias>. If a user attempts
to visit node/<nid> they will receive a 301 Moved Permanently
redirect to /<alias>