Remove cookie_consent reliance on client side
We no longer let the client tell us which IP address it has. We
populate this value on the server side anyway so it makes more
sense to do this when we insert the cookie into the database. This
prevents users changing cookie settings for other IPs.

We now validate the value that is being set for the cookie. This
ensures that random values can't be entered into the database which
could result in a DoS attack.